requirements.txt
pandas
numpy
python-chess
lightgbm
scikit-learn
tqdm
maia2
gdown
pyzstd
mlflow